- Fixing initial config;

+++ src/main/infrastructure/repository/script/init.sh
@@ -11,11 +11,14 @@ DEBUG_OPT=
 trap - INT TERM
 
 # Updates the conf file.
-${DEBUG} && echo "Updating configuration"
-rm -Rf ${PGDATA}/postgresql.conf
-cp /tmp/postgresql.conf ${PGDATA}/postgresql.conf
-rm -Rf ${PGDATA}/pg_hba.conf
-cp /tmp/pg_hba.conf ${PGDATA}/pg_hba.conf
+if [ -f ${PGDATA}/postgresql.conf ]
+then
+	${DEBUG} && echo "Updating configuration"
+	rm -Rf ${PGDATA}/postgresql.conf
+	cp /tmp/postgresql.conf ${PGDATA}/postgresql.conf
+	rm -Rf ${PGDATA}/pg_hba.conf
+	cp /tmp/pg_hba.conf ${PGDATA}/pg_hba.conf
+fi
 
 # Executes the init command.
 exec $@
